+++ src/ViewInstance.js
@@ -4,22 +4,25 @@ import { Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import queryString from 'query-string';
 
-import TitleManager from '@folio/stripes-core/src/components/TitleManager';
-import Pane from '@folio/stripes-components/lib/Pane';
-import PaneMenu from '@folio/stripes-components/lib/PaneMenu';
-import { Accordion, ExpandAllButton } from '@folio/stripes-components/lib/Accordion';
-import KeyValue from '@folio/stripes-components/lib/KeyValue';
-import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
-import Layer from '@folio/stripes-components/lib/Layer';
-import Layout from '@folio/stripes-components/lib/Layout';
-import Button from '@folio/stripes-components/lib/Button';
-import IconButton from '@folio/stripes-components/lib/IconButton';
-import AppIcon from '@folio/stripes-components/lib/AppIcon';
-import Icon from '@folio/stripes-components/lib/Icon';
-import Headline from '@folio/stripes-components/lib/Headline';
-import MultiColumnList from '@folio/stripes-components/lib/MultiColumnList';
-import ViewMetaData from '@folio/stripes-smart-components/lib/ViewMetaData';
-
+import { TitleManager } from '@folio/stripes/core';
+import {
+  Pane,
+  PaneMenu,
+  Row,
+  Col,
+  Button,
+  Accordion,
+  ExpandAllButton,
+  KeyValue,
+  Layer,
+  Layout,
+  IconButton,
+  AppIcon,
+  Icon,
+  Headline,
+  MultiColumnList,
+} from '@folio/stripes/components';
+import { ViewMetaData } from '@folio/stripes/smart-components';
 import { craftLayerUrl } from './utils';
 
 import formatters from './referenceFormatters';
@@ -33,7 +36,6 @@ import ViewMarc from './ViewMarc';
 import makeConnectedInstance from './ConnectedInstance';
 
 const emptyObj = {};
-
 
 class ViewInstance extends React.Component {
   static manifest = Object.freeze({