+++ src/edit/holdings/HoldingsForm.js
@@ -1,18 +1,24 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Paneset from '@folio/stripes-components/lib/Paneset';
-import Pane from '@folio/stripes-components/lib/Pane';
-import PaneMenu from '@folio/stripes-components/lib/PaneMenu';
-import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
-import Button from '@folio/stripes-components/lib/Button';
-import Select from '@folio/stripes-components/lib/Select';
-import TextField from '@folio/stripes-components/lib/TextField';
-import LocationSelection from '@folio/stripes-smart-components/lib/LocationSelection';
-import LocationLookup from '@folio/stripes-smart-components/lib/LocationLookup';
+import {
+  Paneset,
+  Pane,
+  PaneMenu,
+  Row,
+  Col,
+  Button,
+  TextField,
+  Select,
+  ConfirmationModal,
+} from '@folio/stripes/components';
+import {
+  LocationSelection,
+  LocationLookup,
+  ViewMetaData,
+} from '@folio/stripes/smart-components';
 import { Field, FieldArray } from 'redux-form';
-import stripesForm from '@folio/stripes-form';
-import ConfirmationModal from '@folio/stripes-components/lib/ConfirmationModal';
-import ViewMetaData from '@folio/stripes-smart-components/lib/ViewMetaData';
+
+import stripesForm from '@folio/stripes/form';
 
 import renderStatements from './holdingsStatementFields';