+++ src/edit/items/pieceIdentifierFields.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Field } from 'redux-form';
-import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
-import Button from '@folio/stripes-components/lib/Button';
-import TextField from '@folio/stripes-components/lib/TextField';
+import { Row, Col, Button, TextField } from '@folio/stripes/components';
+
 
 const renderPieceIdentifiers = ({ fields, meta: { touched, error, submitFailed }, formatMsg }) => (
   <div>