+++ src/edit/languageFields.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Field } from 'redux-form';
-import Select from '@folio/stripes-components/lib/Select';
+import { Select } from '@folio/stripes/components';
 import RepeatableField from '../components/RepeatableField';
 import languages from '../data/languages';